Add automatic project file changes from Xcode
I did not introduce these changes intentionally. Rather,
I moved a file to a group and then manually reverted the
action.

By adding these changes as a dedicated commit on a dedicated
branch, other developers will not have to deal with them.
Furthermore, it reduces the chance of merge conflicts in case
two developers check in these changes.