Add type-hint for `_relations` in `OsmMapData`
By providing the types, the collection is easier to read
and the code that uses it is easier to understand.

With this commit, I have also updated the enumeration that
seems to have been copy-pasted and that was using "node"
as the parameter name, even though we are dealing with
relations here.